Wu Anti-aliased lines demo, first draft.
I don't think this is ready to go into the `graphics.h` header file but
I wanted to show what I've got so far.

I wrote four "reflections" of the basic Wu algorithm and a wrapper to
call the right one.  (For me it's easier to do this than to figure out a
single version that correctly handles the four cases, and then writing
the single version is easier from the boilerplate.)

I embedded the line-drawing code in a simple demo.

Special-case code for horizontal, vertical and 45° degree lines is
omitted for now.